+++ docs/source/changelog/changelog.md
@@ -4,6 +4,14 @@
 
 **Internal changes & Bugfixes**
 *   Streamlining dependencies
+*   Asset Connection
+	*   [OPC UA] operation - fix problem when InputArguments or OutputArguments node was not present
+	*   [OPC UA] - Use ExpandedNodeId to parse NodeId Strings
+
+**New Features**
+*   Asset Connection
+	*   [OPC UA] - Add ParentNodeId to OpcUaOperationProviderConfig
+	*   [OPC UA] - Introduce mapping between IdShort and Argument Name in OpcUaOperationProviderConfig
 
 ## Release version 0.2.1